requirements.txt
3==0.0.0
absl-py==2.0.0
affine==2.4.0
alabaster==0.7.12
anything==0.2.1
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
astunparse==1.6.3
attrs==23.1.0
Babel==2.9.1
backports.entry-points-selectable==1.1.1
backports.zoneinfo==0.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
binutils==0.2.dev0
BittyTax==0.5.0
but==0.0.1
cachetools==5.3.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cheroot==8.6.0
CherryPy==18.8.0
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.5
contourpy==1.1.1
crispy-bootstrap5==0.7
cryptography==40.0.2
cssselect2==0.7.0
cycler==0.12.1
decorator==5.1.1
distlib==0.3.6
Django==3.2.5
django-bootstrap-datetimepicker==1.5
django-crispy-forms==2.0
django-leaflet==0.28.3
django-paypal==2.0
django-phonenumber-field==7.1.0
django-polymorphic==3.1.0
docutils==0.17.1
ez-setup==0.9
Fabric3==1.14.post1
filelock==3.11.0
fiona==1.9.5
fonttools==4.43.1
GDAL==3.7.2
GeoAlchemy2==0.14.2
geopandas==0.14.0
geoserver-rest==2.5.1
greenlet==3.0.1
html5lib==1.1
idna==3.4
importlib-metadata==6.8.0
importlib-resources==5.4.0
jaraco.classes==3.2.1
jaraco.collections==3.4.0
jaraco.context==4.1.1
jaraco.functools==3.4.0
jaraco.text==3.7.0
Jinja2==3.0.3
jsonschema==3.2.0
kiwisolver==1.4.5
lxml==4.9.2
MarkupSafe==2.0.1
matplotlib==3.8.0
more-itertools==8.14.0
numpy==1.24.4
of==1.0.1
oscrypto==1.3.0
packaging==23.2
pandas==2.1.1
paramiko==2.12.0
paypal==1.2.5
Pillow==9.5.0
pisa==3.0.33
platformdirs==3.2.0
portend==3.0.0
postgis==1.0.4
postgres-helper==0.0.5
psycopg2==2.9.3
psycopg2-binary==2.9.6
pycparser==2.21
Pygments==2.16.1
PyGreSQL==6.0
pyHanko==0.17.2
pyhanko-certvalidator==0.20.1
pymongo==4.1.1
PyNaCl==1.5.0
pyparsing==3.1.1
pypdf==3.8.0
PyPDF2==2.10.9
pypng==0.20220715.0
pyproj==3.6.1
pyrsistent==0.18.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.4.2
reportlab==3.6.12
requests==2.28.2
seaborn==0.13.0
shapely==2.0.2
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.22
sqlparse==0.4.4
svglib==1.5.1
tempora==4.1.2
tinycss2==1.2.1
tqdm==4.64.1
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
uritools==4.0.1
urllib3==1.26.15
virtualenv==20.22.0
webencodings==0.5.1
ws4py==0.5.1
xhtml2pdf==0.2.10
xlrd==1.2.0
XlsxWriter==3.0.3
xmltodict==0.13.0
zc.lockfile==2.0
zipp==3.17.0

